just restructuring a bit
