+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import './App.css';
-import {MainMap} from './map';
+import OpenLayersMap from './components/OpenLayersMap';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <MainMap />
+        <OpenLayersMap />
       </div>
     );
   }
